feat: marketplace api integration
