+++ apps/client/package.json
@@ -11,6 +11,7 @@
         "format": "prettier --write \"src/**/*.{js,jsx,ts,tsx,json,css,md}\""
     },
     "dependencies": {
+        "@repo/shared-dtos": "workspace:*",
         "@repo/ui": "workspace:*",
         "@radix-ui/react-avatar": "^1.1.10",
         "@radix-ui/react-checkbox": "^1.3.2",
