+++ apps/client/src/config/env.ts
@@ -1,0 +1,6 @@
+export const config = {
+    api: {
+        BASE_URL: process.env.API_BASE_URL || 'http://localhost:3000',
+        VERSION: process.env.API_VERSION || 'v1',
+    }
+}
