+++ apps/client/src/lib/constants.ts
@@ -1,0 +1,118 @@
+
+// Sample datas
+// Updated to match Prisma schema with more diverse data
+const rentalPosts = [
+    {
+        posterAddress: '0x2',
+        name: 'Legendary Sword of Azerith',
+        description: "A rare weapon from the NFT game 'Realm of Warriors'.",
+        category: 'Gaming',
+        hourlyRate: 0.05,
+        collateral: 1.5,
+        isBiddable: true,
+        biddingStarttime: '2025-07-18T09:00:00.000Z',
+        biddingEndtime: '2025-07-20T09:00:00.000Z',
+        isActive: true,
+        statusCode: 'AVAILABLE',
+    },
+    {
+        posterAddress: '0x2',
+        name: 'Synthwave Cityscape',
+        description: 'Animated digital art in retro-futuristic style.',
+        category: 'Digital Art',
+        hourlyRate: 0.03,
+        collateral: 0.7,
+        isBiddable: true,
+        biddingStarttime: '2025-07-19T14:00:00.000Z',
+        biddingEndtime: '2025-07-21T14:00:00.000Z',
+        isActive: true,
+        statusCode: 'AVAILABLE',
+    },
+    {
+        posterAddress: '0x2',
+        name: 'Bass Drop #88',
+        description: 'Exclusive EDM loop from top crypto artist SatoshiSounds.',
+        category: 'Music',
+        hourlyRate: 0.02,
+        collateral: 0.5,
+        isBiddable: true,
+        biddingStarttime: '2025-07-19T12:00:00.000Z',
+        biddingEndtime: '2025-07-20T12:00:00.000Z',
+        isActive: true,
+        statusCode: 'AVAILABLE',
+    },
+    {
+        posterAddress: '0x2',
+        name: 'Pixel Wizard Avatar',
+        description: 'Playable avatar for blockchain-based RPG PixelVerse.',
+        category: 'Gaming',
+        hourlyRate: 0.01,
+        collateral: 0.2,
+        isBiddable: false,
+        isActive: true,
+        statusCode: 'AVAILABLE',
+    },
+    {
+        posterAddress: '0x2',
+        name: 'CryptoPunk Poster',
+        description: 'High-res digital art print of CryptoPunk #4587.',
+        category: 'Collectibles',
+        hourlyRate: 0.04,
+        collateral: 0.9,
+        isBiddable: true,
+        biddingStarttime: '2025-07-19T10:30:00.000Z',
+        biddingEndtime: '2025-07-21T10:30:00.000Z',
+        isActive: true,
+        statusCode: 'AVAILABLE',
+    },
+    {
+        posterAddress: '0x2',
+        name: 'Metaverse Fashion Outfit',
+        description: 'A full-body outfit wearable in Decentraland and Sandbox.',
+        category: 'Metaverse',
+        hourlyRate: 0.025,
+        collateral: 0.6,
+        isBiddable: true,
+        biddingStarttime: '2025-07-18T08:00:00.000Z',
+        biddingEndtime: '2025-07-19T08:00:00.000Z',
+        isActive: true,
+        statusCode: 'AVAILABLE',
+    },
+    {
+        posterAddress: '0x2',
+        name: 'Virtual Villa Parcel',
+        description: 'Prime virtual real estate near the MetaBeach.',
+        category: 'Virtual Real Estate',
+        hourlyRate: 0.1,
+        collateral: 2.5,
+        isBiddable: false,
+        isActive: false,
+        statusCode: 'AVAILABLE',
+    },
+    {
+        posterAddress: '0x2',
+        name: 'NFT Basketball Card - Zion Rookie',
+        description: 'Digital card from a limited sports NFT series.',
+        category: 'Sports',
+        hourlyRate: 0.03,
+        collateral: 0.8,
+        isBiddable: true,
+        biddingStarttime: '2025-07-19T15:00:00.000Z',
+        biddingEndtime: '2025-07-20T15:00:00.000Z',
+        isActive: true,
+        statusCode: 'AVAILABLE',
+    },
+    {
+        posterAddress: '0x2',
+        name: 'Crypto Domain: metahub.eth',
+        description: 'Premium ENS domain for metaverse ventures.',
+        category: 'Domain Names',
+        hourlyRate: 0.02,
+        collateral: 0.4,
+        isBiddable: true,
+        biddingStarttime: '2025-07-19T11:00:00.000Z',
+        biddingEndtime: '2025-07-21T11:00:00.000Z',
+        isActive: false,
+        statusCode: 'AVAILABLE',
+    },
+];
