+++ apps/client/tsconfig.json
@@ -1,5 +1,4 @@
 {
-    "extends": "@repo/typescript-config/nextjs.json",
     "compilerOptions": {
         "plugins": [
             {
@@ -8,7 +7,20 @@
         ],
         "paths": {
             "@/*": ["./src/*"]
-        }
+        },
+        "target": "ES2017",
+        "lib": ["dom", "dom.iterable", "esnext"],
+        "allowJs": true,
+        "skipLibCheck": true,
+        "strict": false,
+        "noEmit": true,
+        "incremental": true,
+        "module": "esnext",
+        "esModuleInterop": true,
+        "moduleResolution": "node",
+        "resolveJsonModule": true,
+        "isolatedModules": true,
+        "jsx": "preserve"
     },
     "include": [
         "**/*.ts",
